requirements.txt
streamlit==1.31.1
streamlit-chat==0.1.1
transformers==4.37.2
torch==2.2.0
numpy==1.26.3
google-generativeai==0.8.2
chromadb==0.4.22
python-dotenv==1.0.1
typing-extensions==4.5.0
protobuf==4.25.1
pydantic==2.5.3
tqdm==4.66.1
requests==2.31.0
urllib3==2.0.0